Before login we store the route in vuex and move to this stored route after login. The route is stored in the session storage
